pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "modelfingerprint"
version = "0.1.0"
description = "Standalone file-based model fingerprinting toolkit"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
  "httpx>=0.28,<1.0",
  "jsonschema>=4.23,<5.0",
  "pydantic>=2.11,<3.0",
  "PyYAML>=6.0,<7.0",
  "typer>=0.16,<1.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.18,<2.0",
  "pytest>=8.4,<9.0",
  "ruff>=0.13,<0.14",
  "types-PyYAML>=6.0,<7.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
